implement as part of memop instead
