+++ doc/Syscalls.md
@@ -233,6 +233,5 @@ functionality that is handled by the kernel. `command`, `subscribe`, and
 | 17            | AES              | AES encryption and decryption              |
 | 22            | LPS25HB          | Pressure sensor                            |
 | 154           | Radio            | 15.4 radio interface                       |
-| 254           | System           | Information about running kernel, process  |
 | 255           | IPC              | Inter-process communication                |
