+++ kernel/src/lib.rs
@@ -10,8 +10,8 @@ pub mod debug;
 pub mod driver;
 pub mod ipc;
 pub mod mem;
+pub mod memop;
 pub mod returncode;
-pub mod system;
 pub mod hil;
 
 // Work around https://github.com/rust-lang-nursery/rustfmt/issues/6
