+++ userland/libtock/system.h
@@ -8,8 +8,6 @@ extern "C" {
 
 #define DRIVER_NUM_SYSTEM 254
 
-unsigned system_tock_major_version(void);
-
 void* system_app_memory_begins_at(void);
 void* system_app_memory_ends_at(void);
 void* system_app_flash_begins_at(void);
